Define what the "Compound" class will do. Tests fail at the moment because it's not written.
*  ../periodic_table.py:
Add a getter so that we can look up the elements by symbol in a nicer way.
*  data_for_test.py:
Add a list of all the symbols so we can check that we have them all. Debatable usefulness.
*  test_1.json:
Just to prove that we can load up another table. Again debatable usefulness.
*  test_periodic_table.py:
Add some failing tests for the Compound class that is not implemented yet. Help to define
what it will do.